correceted the list of items for agrihistoregions and cancel nesting intempt. set up style rules manually in historical.xhtml.en A couple of interesting cases for upcoming automatical process: -- How to deal with nested regions (one containing others) -- I tried to rely as much as possible on information in irigaz.owl. In regions.json menu, I labeled Asia_South as South Asia but its label is Asia: South. Not sure what is the most standard labelling. -- Right now, it is only bbox that is a bodyAttribute but I used the bboxes containting the irigaz region. I assume in the future it will be more sophisticated than that -- Interesting to note that in that case, bbox for South_Asia and India is the same -- Note that NE_Brazil are tagged Brazil because I didn't find a better descriptor in the irigaz. -- as well, the South Asia Maprooms are tagged South Asia when they obviously arre desribed by a somewhat random bbox
All the above makes think that probably in the future, irigaz objects will be associated with geometries even though I prentend here that they are associated with a bbox; that Maprooms with non standard bbox will probably be tagged not by an irigaz item but by a bbox. And isPartOf or Contains type of rules will deal with those different relationships.
On second tought, all Maps in Maprooms are a box so maybe they should be always tagged with a bbox, or more generally we need to define standard rules or mechanisms to apply tags and bodyAttributes.